% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2_function.R
\name{plot_postcodes_offencelevel1}
\alias{plot_postcodes_offencelevel1}
\title{Create a plot showing the correlation in a level 1 offence count between two postcodes by month}
\usage{
plot_postcodes_offencelevel1(crime_data, offence_description, postcodes)
}
\arguments{
\item{crime_data}{A data.table object with the following columns:
"date" (POSIXct), "suburb" (chr), "postcode" (chr), "offence_level_1" (chr),
"offence_level_2" (chr), "offence_level_3" (chr), "offence_count" (num).}

\item{offence_description}{A character string of offence description which can be found in
offence level 1. This character string should be UPPERCASE.}

\item{postcodes}{A two-element character vector. Each element is an SA postcode.}
}
\value{
A ggplot object showing the correlation in offence count between the two input postcodes.
}
\description{
\code{plot_postcodes_offencelevel1} takes a data.table object of crime data, a character string
 of description in offence level 1, a two-element character vector of SA postcodes, and creates
 a plot of counts in the input offence for two postcodes by month.
}
\examples{
plot_postcodes_offencelevel1(crime_data, "ACTS INTENDED TO CAUSE INJURY", c(5159, 5173))
}
